refactor: extract memory-book retrieval/indexing into dedicated use-case helper
